4_semester/Операционные системы/fs/03_Cloud/requirement.txt
fuse-python
fastapi
